+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Changed `cargo:rust-env` to use quotes.
+
+### Fixed
+
+- Fixed treating version channel as stable one.
+
 ## [0.0.2] - 2022-12-28
 
 ### Added
@@ -52,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
+[Unreleased]: https://github.com/ferric-bytes/chksum-build/compare/v0.0.2...HEAD
 [0.0.2]: https://github.com/ferric-bytes/chksum-build/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/ferric-bytes/chksum-build/compare/v0.0.0...v0.0.1
 [0.0.0]: https://github.com/ferric-bytes/chksum-build/releases/tag/v0.0.0